+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Asynchronously submit the previous search field and get the next one while processing to reduce network overhead
 - Reuse server connection between claim/submit requests to reduce network overhead
 - Implement higher k-values for the experimental LSD filter, fix a bug and add tests
+- Implement checks for overlap between the most- and least-significant digits
 - Implement a CRT stride filter that allows the process to jump between valid candidates instead of iterating over all of them
 - Configure additional trace logging in the library
 - Replace `--verbose` and `--quiet` with `--log-level`/`-l` and `--no-progress`/`-n`